--- lab02/source/src/figure/compound/square/Rectangle.java
@@ -1,14 +1,0 @@
-package figure.compound.square;
-
-import figure.simple.Point;
-
-public class Rectangle
-{
-	public Point topLeftPoint, bottomRightPoint;
-
-	public Rectangle(Point topLeftPoint, Point bottomRightPoint)
-	{
-		this.topLeftPoint = topLeftPoint;
-		this.bottomRightPoint = bottomRightPoint;
-	}
-}
