--- lab02/source/src/figure/simple/Point.java
@@ -1,13 +1,0 @@
-package figure.simple;
-
-public class Point
-{
-	public float x;
-	public float y;
-
-	public Point(float x, float y)
-	{
-		this.x = x;
-		this.y = y;
-	}
-}
