+++ lab02/src/main/java/com/vlsu/lab02/figure/simple/Point.java
@@ -1,0 +1,13 @@
+package com.vlsu.lab02.figure.simple;
+
+public class Point
+{
+	public float x;
+	public float y;
+
+	public Point(float x, float y)
+	{
+		this.x = x;
+		this.y = y;
+	}
+}
